+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/type/GenerationConfig.kt
@@ -67,7 +67,12 @@ import kotlinx.serialization.Serializable
  * - `application/json`: JSON response in the candidates.
  *
  * @property responseSchema Output schema of the generated candidate text. If set, a compatible
- * [responseMimeType] must also be set.
+ * [responseMimeType] must also be set. This is mutually exclusive with [responseJsonSchema]. Unlike
+ * [responseJsonSchema] this will encode to an OpenAPI schema.
+ *
+ * @property responseJsonSchema Output schema of the generated candidate text. If set, a compatible
+ * [responseMimeType] must also be set. This is mutually exclusive with [responseSchema]. Unlike
+ * [responseSchema] this will encode to an JsonSchema schema, which is the standard moving forward.
  *
  * Compatible MIME types:
  * - `application/json`: Schema for JSON response.
@@ -90,6 +95,7 @@ private constructor(
   internal val stopSequences: List<String>?,
   internal val responseMimeType: String?,
   internal val responseSchema: Schema?,
+  internal val responseJsonSchema: JsonSchema<*>?,
   internal val responseModalities: List<ResponseModality>?,
   internal val thinkingConfig: ThinkingConfig?,
 ) {
@@ -120,6 +126,8 @@ private constructor(
    *
    * @property responseSchema See [GenerationConfig.responseSchema].
    *
+   * @property responseJsonSchema See [GenerationConfig.responseJsonSchema]
+   *
    * @property responseModalities See [GenerationConfig.responseModalities].
    *
    * @see [generationConfig]
@@ -135,8 +143,41 @@ private constructor(
     @JvmField public var stopSequences: List<String>? = null
     @JvmField public var responseMimeType: String? = null
     @JvmField public var responseSchema: Schema? = null
+    @JvmField public var responseJsonSchema: JsonSchema<*>? = null
     @JvmField public var responseModalities: List<ResponseModality>? = null
     @JvmField public var thinkingConfig: ThinkingConfig? = null
+
+    public constructor()
+
+    internal constructor(
+      temperature: Float?,
+      topK: Int?,
+      topP: Float?,
+      candidateCount: Int?,
+      maxOutputTokens: Int?,
+      presencePenalty: Float?,
+      frequencyPenalty: Float?,
+      stopSequences: List<String>?,
+      responseMimeType: String?,
+      responseSchema: Schema?,
+      responseJsonSchema: JsonSchema<*>?,
+      responseModalities: List<ResponseModality>?,
+      thinkingConfig: ThinkingConfig?,
+    ) {
+      this.temperature = temperature
+      this.topK = topK
+      this.topP = topP
+      this.candidateCount = candidateCount
+      this.maxOutputTokens = maxOutputTokens
+      this.stopSequences = stopSequences
+      this.presencePenalty = presencePenalty
+      this.frequencyPenalty = frequencyPenalty
+      this.responseMimeType = responseMimeType
+      this.responseSchema = responseSchema
+      this.responseJsonSchema = responseJsonSchema
+      this.responseModalities = responseModalities
+      this.thinkingConfig = thinkingConfig
+    }
 
     public fun setTemperature(temperature: Float?): Builder = apply {
       this.temperature = temperature
@@ -164,6 +205,9 @@ private constructor(
     public fun setResponseSchema(responseSchema: Schema?): Builder = apply {
       this.responseSchema = responseSchema
     }
+    public fun setResponseSchemaJson(responseSchemaJson: JsonSchema<*>?): Builder = apply {
+      this.responseJsonSchema = responseSchemaJson
+    }
     public fun setResponseModalities(responseModalities: List<ResponseModality>?): Builder = apply {
       this.responseModalities = responseModalities
     }
@@ -172,8 +216,11 @@ private constructor(
     }
 
     /** Create a new [GenerationConfig] with the attached arguments. */
-    public fun build(): GenerationConfig =
-      GenerationConfig(
+    public fun build(): GenerationConfig {
+      if (responseSchema != null && responseJsonSchema != null) {
+        throw InvalidStateException("responseSchema and responseJsonSchema are mutually exclusive.")
+      }
+      return GenerationConfig(
         temperature = temperature,
         topK = topK,
         topP = topP,
@@ -184,10 +231,29 @@ private constructor(
         frequencyPenalty = frequencyPenalty,
         responseMimeType = responseMimeType,
         responseSchema = responseSchema,
+        responseJsonSchema = responseJsonSchema,
         responseModalities = responseModalities,
         thinkingConfig = thinkingConfig
       )
+    }
   }
+
+  public fun toBuilder(): Builder =
+    Builder(
+      temperature = temperature,
+      topK = topK,
+      topP = topP,
+      candidateCount = candidateCount,
+      maxOutputTokens = maxOutputTokens,
+      stopSequences = stopSequences,
+      presencePenalty = presencePenalty,
+      frequencyPenalty = frequencyPenalty,
+      responseMimeType = responseMimeType,
+      responseSchema = responseSchema,
+      responseJsonSchema = responseJsonSchema,
+      responseModalities = responseModalities,
+      thinkingConfig = thinkingConfig
+    )
 
   internal fun toInternal() =
     Internal(
@@ -201,6 +267,7 @@ private constructor(
       presencePenalty = presencePenalty,
       responseMimeType = responseMimeType,
       responseSchema = responseSchema?.toInternalOpenApi(),
+      responseJsonSchema = responseJsonSchema?.toInternalJson(),
       responseModalities = responseModalities?.map { it.toInternal() },
       thinkingConfig = thinkingConfig?.toInternal()
     )
@@ -217,6 +284,7 @@ private constructor(
     @SerialName("presence_penalty") val presencePenalty: Float? = null,
     @SerialName("frequency_penalty") val frequencyPenalty: Float? = null,
     @SerialName("response_schema") val responseSchema: Schema.InternalOpenAPI? = null,
+    @SerialName("response_json_schema") val responseJsonSchema: Schema.InternalJson? = null,
     @SerialName("response_modalities") val responseModalities: List<String>? = null,
     @SerialName("thinking_config") val thinkingConfig: ThinkingConfig.Internal? = null
   )
@@ -229,7 +297,7 @@ private constructor(
      * val config = GenerationConfig.builder()
      * ```
      */
-    public fun builder(): Builder = Builder()
+    @JvmStatic public fun builder(): Builder = Builder()
   }
 }