+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/type/RequestOptions.kt
@@ -27,6 +27,7 @@ internal constructor(
   internal val timeout: Duration,
   internal val endpoint: String = "https://firebasevertexai.googleapis.com",
   internal val apiVersion: String = "v1beta",
+  internal val autoFunctionCallingTurnLimit: Int
 ) {
 
   /**
@@ -38,7 +39,9 @@ internal constructor(
   @JvmOverloads
   public constructor(
     timeoutInMillis: Long = 180.seconds.inWholeMilliseconds,
+    autoFunctionCallingTurnLimit: Int = 10
   ) : this(
     timeout = timeoutInMillis.toDuration(DurationUnit.MILLISECONDS),
+    autoFunctionCallingTurnLimit = autoFunctionCallingTurnLimit
   )
 }