+++ firebase-ai/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- [feature] Added `JsonSchema`, `GenerativeModel#generateObject` and the firebase-ai-ksp processor's
+  annotation (`Generable`, and `Guide`). To learn more about how to use them, and the ksp processor
+  see the [Firebase KSP Processor Readme](https://github.com/firebase/firebase-android-sdk/blob/main/firebase-ai-ksp-processor/README.md)
 - [feature] Added support for configuring thinking levels with Gemini 3 series
   models and onwards. (#7599)
 - [feature] Added support for [API Key